initial jest setup
