+++ package.json
@@ -16,7 +16,7 @@
     "start:build": "tsc -w",
     "start:run": "nodemon dist/index.js",
     "start": "run-p start:**",
-    "test": "echo \"Error: no test specified\" && exit 0"
+    "test": "jest"
   },
   "repository": "github:mwarman/example-rest-node-express-mongo",
   "license": "MIT",
@@ -44,6 +44,7 @@
     "@types/bcrypt": "5.0.0",
     "@types/cors": "2.8.12",
     "@types/express": "4.17.13",
+    "@types/jest": "27.5.1",
     "@types/jsonwebtoken": "8.5.8",
     "@types/morgan": "1.9.3",
     "@types/node": "17.0.31",
@@ -51,14 +52,18 @@
     "@types/passport-anonymous": "1.0.3",
     "@types/passport-http": "0.3.9",
     "@types/passport-jwt": "3.0.6",
+    "@types/supertest": "2.0.12",
     "@typescript-eslint/eslint-plugin": "5.23.0",
     "@typescript-eslint/parser": "5.23.0",
     "eslint": "8.14.0",
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-plugin-import": "2.26.0",
+    "jest": "28.1.0",
     "nodemon": "2.0.15",
     "npm-run-all": "4.1.5",
-    "rimraf": "^3.0.2",
+    "rimraf": "3.0.2",
+    "supertest": "6.2.3",
+    "ts-jest": "28.0.3",
     "typescript": "4.6.4"
   }
 }
